+++ packages/stream_core/lib/stream_core.dart
@@ -7,5 +7,5 @@ export 'src/logger.dart';
 export 'src/platform.dart';
 export 'src/query.dart';
 export 'src/user.dart';
-export 'src/utils.dart' hide SharedEmitterImpl, StateEmitterImpl;
+export 'src/utils.dart' hide SharedEmitterImpl, StateEmitterImpl, objectRuntimeType;
 export 'src/ws.dart';